made shell components
