+++ app/config/routes.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Router, Route, browserHistory, IndexRoute } from 'react-router'
+
+import Main from '../pages/Main'
+import Welcome from '../pages/Welcome'
+
+module.exports = (
+  <Router history={browserHistory}>
+    <Route path="/" component={Main}>
+      <IndexRoute component={Welcome} />
+    </Route>
+  </Router>
+);
