+++ app/reducers/index.js
@@ -1,0 +1,9 @@
+import { combineReducers } from "redux"
+
+import counter from "./counterReducer"
+import user from "./userReducer"
+
+export default combineReducers({
+  counter,
+  user,
+})
