+++ public/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<title>12-Redux-Intro</title>
+</head>
+<body>
+	<div id="app"></div>
+	<script src="bundle.js"></script>
+</body>
+</html>
