+++ webpack.config.js
@@ -1,0 +1,22 @@
+module.exports = {
+  entry: "./app/app.js",
+  output: {
+    filename: "public/bundle.js"
+  },
+  module: {
+    loaders: [
+      {
+        test: /\.js?$/,
+        include: /app/,
+        loader: "babel",
+        query: {
+          presets: ["react", "es2015", "stage-0"],
+          plugins: ["transform-decorators-legacy"],
+        }
+      }
+    ]
+  },
+  // This lets us debug our react code in chrome dev tools. Errors will have lines and file names
+  // Without this the console says all errors are coming from just coming from bundle.js
+  devtool: "eval-source-map"
+};
